pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taskanov"
version = "0.1.0"
description = "TUI for taskmanagement with integrated timetracker, usable with different backends" 
requires-python = ">=3.9"
dependencies = [
    "prompt_toolkit>=3.0.43",
    "google-api-python-client>=2.140.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "python-dateutil>=2.9.0",
    "PyYAML>=6.0.1",
]

[project.scripts]
taskanov = "taskanov.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
